requirements.txt
pyyaml=6.0.1
requests=2.31.0
scikit-learn=1.3.0
tensorflow=2.13.0
pandas=2.0.3
python-dotenv=1.0.0